Parse LocalVarSig
Hardcode StandAloneSig to be LocalVarSig even though it could be StandAloneMethodSig
This will break any parsing that uses a calli instruction for now.

FatFormat has a "Meta Data token" which I'm not certain what it should be...
Set up TableIndex and Signature with generic size (don't implement 4-byte size yet).