Finish implementing all sig, completes TODO(Sig) for now
Well, don't implement FieldMarshal but that's not interesting to me today
All entries in MetadataTables with a type Sig now specify the signature, and use it in NodeValue.

The same heap bytes can be used for MethodRefSig and MethodDefSig, so munge the two types together for reading.

Fix removing [Flags] for ElementType, which was messing up Description.